requirements.txt
python-dotenv
google-generativeai
streamlit
-e .